Remove resetting Vuex state from Profiles.vue
There was still a bug where returning to profile selections screen and
then moving back to manager screen with the same profile (without
changing the active profile in between) displayed an empty local mod
list.

Also if the TsModsModule can't be reset here, it somewhat waters down
the another half of the purpose of these changes, which was to free up
memory when it's not needed. The things that would have gotten reset
are very small compared to the large mod lists.

Full resetting is still done if the active game is changed.