Use fflate to decompress package list chunks
While r2modman used Node's zlib as intended, TSMM used a library that
happened to be present as a 2nd degree dependency. While it worked, the
library had poor performance, taking ~3s to decompress each chunk,
while zlib and fflate takes next to no time at all.

Using DecompressionStream instead was considered, but based on a single
google result it seems that would require updating the TypeScript
version, which is a side quest for another time.

It was also considered sniffing the app version and using fflate only
on TSMM, so it wouldn't need to be added as a dependency on r2modman.
In the end gut feeling decided that would be unnecessarily complex and
prone to break due to future changes.