Option to use Flatpaks under Linux
This adds a configuration setting "linuxUseFlatpak" which controls
whether Steam should be launched as a Flatpak or not. By default this
is detected automatically based on the path, but an option in the
settings can override the behaviour.

If this option is enabled, we first check if we are able to read from
our config directory. If not, we display an error and ask the user to
give us permission. If we do, then we launch Steam through flatpak
rather than the binary directly.