Show progress of each step in the profile import modal
Initially progress was based purely on downloading the mods. It was
recently changed so that extracting the downloaded mod to cache is
allocated 20% of the progress of processing the mod in this step.

However, this still leaves users hanging on 100% completion while the
mods from the cache and configs from the imported profile zip are
copied to the profile folder. These steps now have their own progress
indicators.