Drop .fromReactive() from ThunderstoreMod and ThunderstoreVersion
As far as I can tell, the conversion isn't really needed. The doc
strings of ReactiveObjectConverterInterface claims that access to class
methods is lost, but based on testing this is not true. Another claim
is that this grants access to object's private fields. The fields can
be accessed, but same is true to any ThunderstoreMod object, even if
they aren't passed through Vue's methods. Aside from getting a clear
TypeScript warning about accessing a private field in IDE (compile
time) nothing prevents accessing the field in JS files (run time).

Ideally this might speed up OnlineModList rendering, as the unnecessary
object creation is skipped.