Discard profile if importing fails midway through the process
If importing a profile fails while copying files, the resulting profile
is unlikely to be what the user expects it to be, so delete the profile
folder. The process can't be resumed, so any retry will do all the work
(besides downloading the mods) again anyway.

When creating a new profile, it's the newly created profile folder.
When updating an existing profile, it's the temporary work folder.

On the UI/state management side, we need to check if the profile
exists or not, and update the profile list accordingly.