Drop remnants of concurrent profile imports
Originally user could close the profile import progress modal and start
importing another profile. Option to close the modal was removed when
it was detected that this lead to weird behaviour when active profile
was changed while an import was in progress. Some of the code related
to this feature was left behind in the hopes that it could later be
fixed and brought back.

ImmutableProfiles were introduced, and it in fact seems they fix the
immediate issues with changing the profile. However, I decided the
feature won't be coming back anytime soon based on the following:

- It seems unlikely to me that user would want to import multiple
  profiles at the same time. I've no data to back up this claim
- Users currently have no way of knowing when an import finishes once
  they close the modal (unlike regular downloads, which have a
  dedicated view for this), which may cause confusion
- Having multiple concurrent imports seems to degrade the performance
  of the app
- Things still hard break if the active game is changed while the
  import is in progress. This could probably be fixed by finishing
  rolling out the ImmutableProfile to all parts of the code base

Even if the feature is brought back some day, the component is now so
heavily refactored that it's easier to start the implementation from
the scratch than based on the old one.