Return combos in the order based on an argument
When ThunderstoreCombos were created from data fetched from IndexedDB,
the objects were ordered in which ever order the database returned
them. Sort the combos based on the dependency string array argument
instead to guarantee consistent ordering.

At least for profile imports this is meaningful, as the mods will be
shown in this order when "custom" sort order is selected. Another place
where this might be an improvement is the buildDependencySet helper in
the download providers.