Reduce time spent on the splash screen
Users may want to get to the manager screen quickly so they can just
launch the game with existing mods. Having to wait for the latest mod
list update can be frustrating.

Splash screen now only checks the API for a fresh list if no local
cache is found at all. Handling different error scenarios is also
omitted, and user is moved to profile selection screen even if no mod
list can be loaded. Error screen offering a chance to retry loading
the mod list is dropped. The idea is that the background update or a
manually triggered mod list update will handle the mod list eventually.

Note that the changes only affect the splash screen. Background updates
still work the same way they did before.

In the future the splash screen can be hastened even further e.g. by
handling the loading of the exclusion list in the background and
reducing the timeouts/retry attempts of the API calls.