Add Hook, Line & Sinker to exclusions json.
HLS is a mod manager itself and should thus be excluded from the r2 mod
manager. It was already added to .md file (which is used by all the
versions of the mod manager to check the latest changes), but not to
the .json file (which is bundled with the build and used as a fallback
if user is offline).