Implement enabling/disabling mods via PackageInstaller implementations
PackageInstallers can now define custom methods for disabling and
enabling mods instead of depending on the legacy "one-size fits all"
approach.

The support is added only for regular packages. Mod loaders will not
implement disabling/enabling, at least at this time, as it's seen a
difficult task to define properly across the board. For the matter,
currently disabling mod loader packages seems to do little, as the
game still loads the mods if it's launched modded.

Also provide helpers for disabling/enabling mods by appending the files
with ".old" extension. This is in line with how
ProfileInstallerProvider.applyModModeForSubdir handles things.