Add ModListUpdateBanner
The banner will show on top of an empty online mod list, explaining
that the list is empty because splash screen failed to load it. A
one-click option for trying to update the mod list is offered.

This change is adjustment to the fact that the splash screen no longer
stops on errors and user is more likely to reach the manager view with
no mod list loaded into Vuex.