Fix an issue with exclusion list not getting updated when needed
When updating the persistent cache, a sanity check ensured the
exclusion list loaded before trying to use it. However, as an empty
array was used as the default value (and the value used when the store
was reset), the sanity check never triggered.

It's unlikely this has caused issues before, as the exclusion list
should've always been updated before updating the persistent cache.
This however changed when Splash screen was changed to skip unnecessary
steps to get the user to profile selection screen faster.