Extend the argument types accepted by Vuex methods
Some call sites of the Vuex methods pass objects of wrong but
compatible type to the methods. TypeScript doesn't pick up on this as
the type support for the old Vuex version is non-existent. Extend the
argument types so some later changes to methods don't break when the
ExportMod object is passed.