Combine preview modals when importing a profile
The modal now shows a list of the known mods which can be installed, as
well as a list of the names of the unknown mods. If there's no known
mods, the import button is disabled. Option to cancel the import is
always shown, in case user wants to cancel after seeing the contents of
the profile.

The main reason for combining the modals is to make it easier to
provide an option for updating the mod list without leaving the modal.
This became important after the recent changes, as the splash screen no
longer guarantees that the mod list is loaded into Vuex.