Rename Vuex store attributes and methods
The "background" update originally referred to the periodic updates,
i.e. not the update that happens on splash screen and shows the
progress to user. The Vuex state is now shared between the background
updates and user triggered manual updates that show a lightweight
status reporting, so the naming was bit misleading.