Move refreshThunderstoreModList to Vuex
As the number of callsites increases, it makes more sense to have the
method for downloading the latest mod list, storing it into IndexedDB
and loading packages to Vuex inside Vuex, so each component doesn't
need to extend the UtilityMixin just to access it.

For TSMM this also allows injecting telemetry gathering to the method
inside Vuex, and not needing to due it awkwardly in App.vue.