Turn off autocomplete for text inputs
TSMM updates to use Chrome v131, which causes awkward autocomplete
popups to be shown for text inputs. This seemed to affect only inputs
with ids, but not all of them. To avoid regressions if ids would be
added later, they were added now, paired with the autocomplete="off"
attribute. As an additional benefit, this makes it easier to interpret
errors TSMM logs into Sentry.