Fix state drift of ignoreCache setting
Toggling the setting didn't take place immediately everywhere.
Assumedly this is due to Vuex not detecting changes in a class based
object.

To fix this, the setting is encapsulated completely in Vuex: the value
is initialized when settings are loaded, it's changed only through a
Vuex action, and it's always read from Vuex state. The only exception
to this is the ManagerSettingsMigration but I assume this is such a
niche case it's not worth addressing.