Adjust CdnProvider to use more complicated data structure
Thunderstore is migrating to another CDN provider, and the change is to
be done as a gradual rollout. To avoid requiring a new release for each
rollout change, the mod manager will be changed to read the CDN
configuration from an online source.

The requirements of this task requires a more complicated data
structure than the simple array used previously. Firstly we want to
control the percentage of requests that are sent to old main CDN and
the new main CDN. Secondly we want to test the change with only package
downloads first, meaning other CDN dependant operations should only use
the old main and mirror CDNs. Thirdly when the mirror CDN is selected,
portion of the requests are not directed to the new CDN so users have a
way to opt out should problems arise.

It should be noted the implementation was kept as simple as it could
given the requirements above. Therefore it might make sense to revisit
the implementation once the rollout is complete, all operations can use
the new CDN, and the requirements may have changed.