Use latest ecosystem data from Thunderstore API
Drop the hardcoded ecosystem data and schema files used in prototyping
phase. Instead, use the files that can be refreshed with the sync
script included in the repo.

Use the TypeScript types generated based on the JSON Schema to make the
typing more exact and e.g. allows dropping some type casting elsewhere
on the codebase. This also improves the perceived performance in IDE
since the types inferred based on the data could contain thousands of
alternative types for some objects.

The quicktype library that's used to generate the types could in theory
be used to also validate the data via the Convert class available in
the type file. However, the error messages it provides are completely
useless in debugging the problem, so ajv library is still used even
though it adds another dependency.