Make getFullDependencyList act more like previous dependency resolution
There isn't a clear, logical spec for how dependency resolution should
work. While switching to use only one method to download mods, the
dependency resolution was also unified under one helper, but this
inadvertently brought in a subtle change.

Changes in this commit brings the implementation closer to the original
one. Namely dependencies that already have some version installed in
the profile are ignored only when installing a specific version of a
single mod, instead of doing so whenever USE_LATEST_VERSION is used.
In other words, "update all" no longer does this filtering. Instead,
the assumption is that all installed mods, including dependencies, are
updated at the same time and are included in the `combos` parameter.

This still leaves the difference where duplicates of mods included in
the `combos` are filtered out, but I think this is more a bug fix than
a change, as only effect I can think it having is enforcing the
requested version (included in `combos`) is installed in case there's
a circular dependency.