Drop Dexie mocking from DependencyUtils tests
Dexie can't be used in unit tests as the underlying IndexedDB isn't
available on Node environment where the tests run. Original attempt to
mock it bled to ModLinker unit tests (which use SettingsDexieStore)
despite the attempts to unmock the implementation after DependencyUtils
tests.

As an alternative approach, mock only the line of the helper function
participating in the tests that actuallu calls Dexie. To do this, the
line had to be placed in a separate file so that the helper function in
PackageDexieStore calls the mocked version (having the line on the same
file would cause PackageDexieStore to call unmocked version). The
complexity of the additional file can be justified by the fact that
this way we don't have to mock the whole helper function, which could
lead to bugs in unit tests.

Also does some minor cleanup on the unit test file to reduce noise and
make it easier to focus on what the tests actually do.