Use PackageLoader enum provided by Thunderstore ecosystem
Drop an extra step from mapping PackageLoader to package loader/plugin
installer. While installers probably shouldn't be bound to the loader,
at least for plugins, at least the approach is not a bit simpler.

Potential future work includes further simplifying the mapping by
adding a "NoOpInstaller" that can be returned for PackageLoader.None,
and a "LegacyPluginInstaller" that wraps the old InstallRuleInstaller
in PackageInstaller interface for PackageLoaders that currently return
null from getPluginInstaller. This is out of the current scope as it
would require changes to GenericProfileInstaller. Furthermore, the
changes would become obsolete when TCLI handles the installations.