Update latest data from ecosystem and cast enum keys to UPPER_CASE
To avoid having to change all the usages of enums now controlled by
Thunderstore ecosystem, use regexp to cast the enum keys to the format
used by the mod manager. We need to resort to regexp as the ecosystem
uses kebab-case and we can't control how Quicktype generates the keys
from the values.