Simplify download progress tracking
Axios reports download progress in bytes. Previously this was casted to
percentages to report the download progress outside the download
provider. Now download provider also reports the progress in bytes and
lets UI handle casting to percentages if it so chooses. This lead into
a situation where the download provider unnecessarily casts the bytes
to percentages to bytes again.