Add enable/disable support to InstallRulePluginInstaller
ReturnOfModdingPluginInstaller also depends on this implementation, as
it's built on top of the IRPI.

Some extra logic is still left in GenericProfileInstaller - these can
be cleaned out once getPluginInstaller starts returning IRPI as the
installer for the games that currently fall back to legacy
implementation that's embedded into GenericProfileInstaller.

Other than that, GenericProfileInstaller no longer has any enable or
disable logic - it's all compartmentalized inside the various classes
that implement PackageInstaller.