Move mod uninstallation related code from GenericProfileInstaller
This is a step on the path to make InstallRulePluginInstaller support
all four methods of PackageInstaller class.

I'm not sure if it matters in which order uninstallState and
uninstallModLoader is called but decide to keep the order unchanged.

Likewise uninstallModLoader behaves the same way it used to, i.e.
even if mod loader files are uninstalled, the rest of the non-loader
mod uninstallation code is still executed.