Include settingsIdentifier in Error when saving game settings fail
This is to shed some light to the error "Failed to execute 'put' on
'IDBObjectStore': Evaluating the object store's key path yielded a
value that is not a valid key." that's seen in production.

The problem might be related to changes done when the mod manager was
changed to read the game configs from Thunderstore ecosystem schema.
Or it might be caused by users who mess up their Vuex store state by
using mouse to navigate back in history, although I think that should
cause a different error, but it might be eater up by IndexedDB.