Simplify LaunchType enum related type handling
It seems the code was jumping through extra hoops that might be
required when dealing with integer enums, but are unnecessary for
string based enums like LaunchType.

Things could perhaps be further simplified if LaunchType was used
instead of string when typing SettingsDexieStore, and/or if it would be
okay for ManagerSettings to return "AUTO" when the value is undefined.