Extract package loader uninstall logic from GenericProfileInstaller
Import it into package loader installers that don't have a custom
implementation. Now all package loader installers implement uninstall,
but this doesn't apply to plugin installers yet.

It's worth noting that this duplicates the implementation, as the
package loader uninstall logic is still kept in GenericProfileInstaller
too, since the version extracted to PackageInstaller handles errors
slightly differently. The verison in GenericProfileInstaller will be
removed later as the refactoring process progresses.