Add missing methods to LovelyPluginInstaller
The custom installation method tracks intallations in a state file. Use
the existing state-based methods to uninstall and disable/enable mods.

This subtly changes the old behaviour since uninstallSubDir and
applyModeSubDirs are no longer called like they were in the
implementation that resided in GenericProfileInstaller. I assume it's
safe to ignore this change and that the state-based methods contain
all the logic required for mods installed with addToStateFile.

It's out of the current scope to figure out if lovely actually needs a
custom plugin installer or if it would do just fine with the
InstallRulePluginInstaller, which has evolved since support for Lovely
as introduced.