Make PackageInstaller.uninstall a required method
All PackageInstaller implementations, for both mod loaders and plugins,
now implement the method so it can be required to ensure future
implementations don't skip it.

Disable/Enable methods are implemented for all plugin installers as
well, but the operation is not supported for mod loader packages, so
the methods remain optional in the interface.