DO NOT MERGE: Manually edit ecosystem data to include Broforce
Normally these files are updated by reading the latest data from the
Thunderstore API using the sync command in the repo. Here they are
edited manually for development purposes. This commit should be dropped
from the real PR, which should be merged after the real data is already
merged.

In ecosystem.json, packageIndex field points to thunderstore.dev
environment. You should be able to upload the UMM package and test mods
to https://thunderstore.dev/c/test/ and the mod manager should pick
them up. Note that due to caching this might take up to 3h.

At the end of ecosystem.json, the mod loader package is registered.
This is required so the mod manager knows it's a mod loader and is
installed differently from the regular mods.

- packageId is teamName-packageName of the mod loader package. It
  doesn't have to be UMM-UMM, as long as it matches the name that is
  registered in ecosystem.json. It's good practice to use a separate
  team to manage the mod loader package, so new contributors can be
  added to the team without giving them access to your personal mods
- rootFolder refers to a folder inside the package zip. It should
  contain the mod loader files. It doesn't have to "UMM".
- loader is just a identifier. It should match packageLoader field
  in Broforce's entry earlier in the file. Again, it doesn't have to
  be "umm".